+++ R/data.R
@@ -53,6 +53,25 @@
 
 
 
+#' Efficiency data frame column names
+#'
+#' A string list containing named names of columns in efficiency data frames.
+#' Items in the list provide default values for column name function arguments
+#' to efficiency functions throughout the `Recca` package.
+#'
+#' @format A string list with `r length(efficiency_cols)` entries.
+#' \describe{
+#' \item{eta_i}{The name of a column in a wide-by-matrices data frame containing industry efficiencies.}
+#' \item{eta_pfd_gross}{The name of a column in a wide-by-matrices data frame containing  efficiencies between the primary and gross final demand stages.}
+#' \item{eta_pfd_net}{The name of a column in a wide-by-matrices data frame containing efficiencies between the primary and net final demand stages.}
+#' }
+#'
+#' @examples
+#' efficiency_cols
+"efficiency_cols"
+
+
+
 #' Sankey diagram data frame column names
 #'
 #' A string list containing named names of columns in Sankey data frames.