+++ data-raw/create_constants.R
@@ -36,10 +36,20 @@ usethis::use_data(psut_cols, overwrite = TRUE)
 #
 
 aggregate_cols <- list(aggregate_primary = "EX.p",
-                       net_aggregate_demand = "EX.d_net",
-                       gross_aggregate_demand = "EX.d_gross",
+                       net_aggregate_demand = "EX.fd_net",
+                       gross_aggregate_demand = "EX.fd_gross",
                        region = "Region")
 usethis::use_data(aggregate_cols, overwrite = TRUE)
+
+
+#
+# Default names for columns in efficiency data frames
+#
+
+efficiency_cols <- list(eta_i = "eta_i",
+                        eta_pfd_gross = "eta_pfd_gross",
+                        eta_pfd_net = "eta_pfd_net")
+usethis::use_data(efficiency_cols, overwrite = TRUE)
 
 
 #